Implemented Median function
Determine if the length of the array is a even or odd length. If odd subtract 1
from index length and divide by two to determine the middle index (median). If even
then just divide the index length by two. Mark length var to determine the
correct return value. If odd return the determined indexed value, if even then
add the two mid values and divide by two to get the median of values.

- Altered the 2nd test for median since the check was incorrect. The median of
  0.0, 0.5, -1.0, 1.0 is 0.25 not 0.0.