Stand. Dev. funtion implemented
Ultiizing the Standard Deviation algorithm from the URL included we first
determine the mean and length of the array. Then we subtract each value in the
array from the mean and square it. Using the sum of those values divided by our
length we get our varience. Applying square root to the variance gives us the
stand. dev.