L2 Norm function implemented
The Euclidean norm or L2 norm is calculated by taking the square root of a
a sum of squared numbers in a set. Used a for loop to determine the sum of each
array value squared, then i sqrt() that result. Made a simple check to determine
if the list is empty for the match block for the return options.